extract ContrstAdminController
